# swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Hotel API",
        "description": "API for hotel management",
        "version": "1.0"
    },
    "produces": [
        "application/json"
    ],
    "host": "127.0.0.1:3000",
    "basePath": "/",
    "paths": {
...
        "/clients": {
            "get": {
                "operationId": "clients",
                "tags": [
                    "Clients"
                ],
                "description": "Récupère la liste des clients",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object",
                                "properties": {
                                    "id_client": {
                                        "type": "integer"
                                    },
                                    "first_name": {
                                        "type": "string"
                                    },
                                    "last_name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "phone": {
                                        "type": "string"
                                    },
                                    "registration_date": {
                                        "type": "string",
                                        "format": "date"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
...
        "/reservations": {
            "get": {
                "operationId": "reservations",
                "tags": [
                    "Reservations"
                ],
                "description": "Récupère la liste des reservations",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object",
                                "properties": {
                                    "id_reservation": {
                                        "type": "integer"
                                    },
                                    "id_client": {
                                        "type": "integer"
                                    },
                                    "id_room": {
                                        "type": "integer"
                                    },
                                    "check_in_date": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "check_out_date": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "total_cost": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "reservation_status": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
...
        "/payments": {
            "get": {
                "operationId": "payments",
                "tags": [
                    "Payments"
                ],
                "description": "Récupère la liste des paiements",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object",
                                "properties": {
                                    "id_payment": {
                                        "type": "integer"
                                    },
                                    "payment_date": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "amount": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "payment_method": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
...
    }
}